% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-vscode.R
\name{TapReporter}
\alias{TapReporter}
\title{Test reporter: TAP format.}
\description{
This reporter will output results in the Test Anything Protocol (TAP),
a simple text-based interface between testing modules in a test harness.
For more information about TAP, see http://testanything.org
}
\concept{reporters}
\section{Super class}{
\code{\link[testthat:Reporter]{testthat::Reporter}} -> \code{TapReporter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TapReporter$new()}}
\item \href{#method-start_context}{\code{TapReporter$start_context()}}
\item \href{#method-add_result}{\code{TapReporter$add_result()}}
\item \href{#method-end_reporter}{\code{TapReporter$end_reporter()}}
\item \href{#method-clone}{\code{TapReporter$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".start_context">}\href{../../testthat/html/Reporter.html#method-.start_context}{\code{testthat::Reporter$.start_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_line">}\href{../../testthat/html/Reporter.html#method-cat_line}{\code{testthat::Reporter$cat_line()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_tight">}\href{../../testthat/html/Reporter.html#method-cat_tight}{\code{testthat::Reporter$cat_tight()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_context">}\href{../../testthat/html/Reporter.html#method-end_context}{\code{testthat::Reporter$end_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_context_if_started">}\href{../../testthat/html/Reporter.html#method-end_context_if_started}{\code{testthat::Reporter$end_context_if_started()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_file">}\href{../../testthat/html/Reporter.html#method-end_file}{\code{testthat::Reporter$end_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_test">}\href{../../testthat/html/Reporter.html#method-end_test}{\code{testthat::Reporter$end_test()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="is_full">}\href{../../testthat/html/Reporter.html#method-is_full}{\code{testthat::Reporter$is_full()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="local_user_output">}\href{../../testthat/html/Reporter.html#method-local_user_output}{\code{testthat::Reporter$local_user_output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="rule">}\href{../../testthat/html/Reporter.html#method-rule}{\code{testthat::Reporter$rule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_file">}\href{../../testthat/html/Reporter.html#method-start_file}{\code{testthat::Reporter$start_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_reporter">}\href{../../testthat/html/Reporter.html#method-start_reporter}{\code{testthat::Reporter$start_reporter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_test">}\href{../../testthat/html/Reporter.html#method-start_test}{\code{testthat::Reporter$start_test()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="update">}\href{../../testthat/html/Reporter.html#method-update}{\code{testthat::Reporter$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TapReporter$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_context"></a>}}
\if{latex}{\out{\hypertarget{method-start_context}{}}}
\subsection{Method \code{start_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TapReporter$start_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_result"></a>}}
\if{latex}{\out{\hypertarget{method-add_result}{}}}
\subsection{Method \code{add_result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TapReporter$add_result(context, test, result)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_reporter"></a>}}
\if{latex}{\out{\hypertarget{method-end_reporter}{}}}
\subsection{Method \code{end_reporter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TapReporter$end_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TapReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
